Restructure unit tests into folders
