+++ tests/kinematics/kinematics_strategies.py
