+++ tests/vision/test_bounding_box.py
