+++ tests/vision/test_images.py
