+++ tests/vision/test_vision_utils.py
