+++ tests/vision/vision_strategies.py
@@ -9,7 +9,7 @@ from hypothesis.extra import numpy as numpy_st
 from robotics_utils.vision.bounding_box import BoundingBox
 from robotics_utils.vision.images import DepthImage, Image, PixelXY, RGBImage
 
-from .common_strategies import integer_ranges
+from ..common_strategies import integer_ranges
 
 
 @st.composite
